Split type checker mod.rs into diagnostics, solving, and registration modules
