+++ src/check/mod.rs
@@ -11,21 +11,27 @@
 ///   Pass 3: Substitute solved types into expr_types (mod.rs)
 ///
 /// Split into:
-///   mod.rs    — Checker struct, public API, solving, registration
-///   types.rs  — TyVarId, Constraint, resolve_vars
-///   infer.rs  — Expression/statement inference
-///   calls.rs  — Function call resolution
+///   mod.rs          — Checker struct, public API, declaration checking
+///   types.rs        — TyVarId, Constraint, resolve_vars
+///   infer.rs        — Expression/statement inference
+///   calls.rs        — Function call resolution
+///   registration.rs — Function/type/protocol declaration registration
+///   solving.rs      — Constraint solving (unification)
+///   diagnostics.rs  — Error hint helpers
 
 mod types;
 mod infer;
 pub(crate) mod calls;
+mod registration;
+mod solving;
+mod diagnostics;
 
 use std::collections::HashMap;
 use crate::ast;
 use crate::ast::ExprId;
 use crate::diagnostic::Diagnostic;
-use crate::types::{Ty, TypeEnv, FnSig, VariantCase, VariantPayload, ProtocolDef, ProtocolMethodSig};
-use types::{TyVarId, Constraint, is_inference_var, UnionFind, resolve_ty};
+use crate::types::{Ty, TypeEnv, VariantCase, VariantPayload, ProtocolDef, ProtocolMethodSig};
+use types::{TyVarId, Constraint, UnionFind, resolve_ty};
 
 pub(crate) fn err(msg: impl Into<String>, hint: impl Into<String>, ctx: impl Into<String>) -> Diagnostic {
     Diagnostic::error(msg, hint, ctx)
@@ -253,378 +259,6 @@ impl Checker {
         let module_types = std::mem::take(&mut self.expr_types);
         self.expr_types = saved.0; self.infer_types = saved.1; self.constraints = saved.2; self.uf = saved.3;
         module_types
-    }
-
-    // ── Constraint solving ──
-
-    fn solve_constraints(&mut self) {
-        let constraints = std::mem::take(&mut self.constraints);
-        // Union-Find makes constraint solving order-independent.
-        // A single pass suffices; re-processing is a harmless no-op.
-        for c in &constraints {
-            self.unify_infer(&c.expected, &c.actual);
-        }
-        // Emit diagnostics for unresolvable mismatches
-        for c in &constraints {
-            if !self.unify_infer(&c.expected, &c.actual) {
-                let exp = resolve_ty(&c.expected, &self.uf);
-                let act = resolve_ty(&c.actual, &self.uf);
-                if exp != Ty::Unknown && act != Ty::Unknown {
-                    let hint = Self::hint_with_conversion(
-                        "Fix the expression type or change the expected type",
-                        &exp, &act,
-                    );
-                    self.emit(err(
-                        format!("type mismatch in {}: expected {} but got {}", c.context, exp.display(), act.display()),
-                        hint, c.context.clone()).with_code("E001"));
-                }
-            }
-        }
-    }
-
-    pub(crate) fn suggest_conversion(expected: &Ty, actual: &Ty) -> Option<String> {
-        match (actual, expected) {
-            (Ty::Int, Ty::String) => Some("use `int.to_string(x)` to convert Int to String".to_string()),
-            (Ty::Float, Ty::String) => Some("use `float.to_string(x)` to convert Float to String".to_string()),
-            (Ty::Bool, Ty::String) => Some("use `to_string(x)` to convert Bool to String".to_string()),
-            (Ty::String, Ty::Int) => Some("use `int.parse(s)` to convert String to Int (returns Result[Int, String])".to_string()),
-            (Ty::String, Ty::Float) => Some("use `float.parse(s)` to convert String to Float (returns Result[Float, String])".to_string()),
-            (Ty::Float, Ty::Int) => Some("use `to_int(x)` to convert Float to Int (truncates)".to_string()),
-            (Ty::Int, Ty::Float) => Some("use `to_float(x)` to convert Int to Float".to_string()),
-            _ => None,
-        }
-    }
-
-    pub(crate) fn hint_with_conversion(base_hint: &str, expected: &Ty, actual: &Ty) -> String {
-        if let Some(conv) = Self::suggest_conversion(expected, actual) {
-            format!("{}. Or {}", base_hint, conv)
-        } else {
-            base_hint.to_string()
-        }
-    }
-
-    fn unify_infer(&mut self, a: &Ty, b: &Ty) -> bool {
-        // Inference vars: union/bind immediately, always succeeds.
-        // Conflicting concrete bindings are unified structurally when possible,
-        // but the inference var case never returns false — matching HashMap semantics.
-        match (is_inference_var(a), is_inference_var(b)) {
-            (Some(ia), Some(ib)) => {
-                self.uf.union(ia.0, ib.0);
-                true
-            }
-            (Some(ia), None) => {
-                let b_resolved = resolve_ty(b, &self.uf);
-                if !self.uf.occurs(ia.0, &b_resolved) {
-                    if let Some(existing) = self.uf.bind(ia.0, b_resolved.clone()) {
-                        // Existing binding — try structural unify but don't fail
-                        self.unify_infer(&existing, &b_resolved);
-                    }
-                }
-                true
-            }
-            (None, Some(ib)) => {
-                let a_resolved = resolve_ty(a, &self.uf);
-                if !self.uf.occurs(ib.0, &a_resolved) {
-                    if let Some(existing) = self.uf.bind(ib.0, a_resolved.clone()) {
-                        self.unify_infer(&a_resolved, &existing);
-                    }
-                }
-                true
-            }
-            (None, None) => {
-                let a_resolved = resolve_ty(a, &self.uf);
-                let b_resolved = resolve_ty(b, &self.uf);
-                self.unify_structural(&a_resolved, &b_resolved)
-            }
-        }
-    }
-
-    fn unify_structural(&mut self, a: &Ty, b: &Ty) -> bool {
-        if *a == Ty::Unknown || *b == Ty::Unknown { return true; }
-        match (a, b) {
-            (Ty::Applied(id1, args1), Ty::Applied(id2, args2)) if id1 == id2 && args1.len() == args2.len() => {
-                args1.iter().zip(args2.iter()).all(|(x, y)| self.unify_infer(x, y))
-            }
-            (Ty::Tuple(a), Ty::Tuple(b)) if a.len() == b.len() =>
-                a.iter().zip(b.iter()).all(|(x, y)| self.unify_infer(x, y)),
-            (Ty::Fn { params: ap, ret: ar }, Ty::Fn { params: bp, ret: br }) if ap.len() == bp.len() =>
-                ap.iter().zip(bp.iter()).all(|(x, y)| self.unify_infer(x, y)) && self.unify_infer(ar, br),
-            (Ty::Record { fields: fa }, Ty::Record { fields: fb }) => {
-                fa.len() == fb.len() && fa.iter().all(|(n, t)| fb.iter().any(|(n2, t2)| n == n2 && self.unify_infer(t, t2)))
-            }
-            (Ty::OpenRecord { fields: req, .. }, Ty::Record { fields: actual })
-            | (Ty::OpenRecord { fields: req, .. }, Ty::OpenRecord { fields: actual, .. }) => {
-                req.iter().all(|(n, t)| actual.iter().any(|(n2, t2)| n == n2 && self.unify_infer(t, t2)))
-            }
-            (Ty::Named(na, args_a), Ty::Named(nb, args_b)) if na == nb => {
-                args_a.len() == args_b.len()
-                    && args_a.iter().zip(args_b.iter()).all(|(ta, tb)| self.unify_infer(ta, tb))
-                    || (args_a.is_empty() || args_b.is_empty())
-            }
-            (Ty::Named(_, _), _) => {
-                let resolved = self.env.resolve_named(a);
-                if resolved != *a { self.unify_infer(&resolved, b) } else { a.compatible(b) }
-            }
-            (_, Ty::Named(_, _)) => {
-                let resolved = self.env.resolve_named(b);
-                if resolved != *b { self.unify_infer(a, &resolved) } else { a.compatible(b) }
-            }
-            _ => a.compatible(b),
-        }
-    }
-
-    // ── Registration ──
-
-    /// Collect structural bounds from generic params: Record → OpenRecord conversion.
-    fn collect_structural_bounds(&self, generics: &Option<Vec<ast::GenericParam>>) -> HashMap<String, Ty> {
-        let mut sb = HashMap::new();
-        let gs = match generics { Some(gs) => gs, None => return sb };
-        for g in gs {
-            let bte = match &g.structural_bound { Some(bte) => bte, None => continue };
-            let bt = self.resolve_type_expr(bte);
-            sb.insert(g.name.clone(), match bt { Ty::Record { fields } => Ty::OpenRecord { fields }, o => o });
-        }
-        sb
-    }
-
-    /// Build a prefixed key: "module.name" or just "name".
-    fn prefixed_key(prefix: Option<&str>, name: &str) -> String {
-        prefix.map(|p| format!("{}.{}", p, name)).unwrap_or_else(|| name.to_string())
-    }
-
-    /// Collect protocol bounds from generic params: TypeVar name → list of protocol names.
-    fn collect_protocol_bounds(&self, generics: &Option<Vec<ast::GenericParam>>) -> HashMap<String, Vec<String>> {
-        let mut pb = HashMap::new();
-        let gs = match generics { Some(gs) => gs, None => return pb };
-        for g in gs {
-            if let Some(bounds) = &g.bounds {
-                if !bounds.is_empty() {
-                    pb.insert(g.name.clone(), bounds.clone());
-                }
-            }
-        }
-        pb
-    }
-
-    fn register_fn_sig(&mut self, name: &str, params: &[ast::Param], return_type: &ast::TypeExpr,
-                        effect: &Option<bool>, r#async: &Option<bool>, generics: &Option<Vec<ast::GenericParam>>, prefix: Option<&str>) {
-        let gnames: Vec<String> = generics.as_ref().map(|gs| gs.iter().map(|g| g.name.clone()).collect()).unwrap_or_default();
-        let sb = self.collect_structural_bounds(generics);
-        let pb = self.collect_protocol_bounds(generics);
-        for gn in &gnames { self.env.types.insert(gn.clone(), Ty::TypeVar(gn.clone())); }
-        let ptys: Vec<(String, Ty)> = params.iter().map(|p| (p.name.clone(), self.resolve_type_expr(&p.ty))).collect();
-        let ret = self.resolve_type_expr(return_type);
-        for gn in &gnames { self.env.types.remove(gn); }
-        let is_effect = effect.unwrap_or(false) || r#async.unwrap_or(false);
-        let key = Self::prefixed_key(prefix, name);
-        if prefix.is_none() && is_effect { self.env.effect_fns.insert(name.to_string()); }
-        let min_p = params.iter().take_while(|p| p.default.is_none()).count();
-        self.env.functions.insert(key.clone(), FnSig { params: ptys, ret, is_effect, generics: gnames, structural_bounds: sb, protocol_bounds: pb });
-        if min_p < params.len() {
-            self.env.fn_min_params.insert(key, min_p);
-        }
-    }
-
-    fn validate_protocols(&mut self, derives: &[String], type_name: &str) {
-        for d in derives {
-            if !self.env.protocols.contains_key(d.as_str()) {
-                let valid: Vec<&str> = self.env.protocols.keys().map(|s| s.as_str()).collect();
-                self.emit(err(
-                    format!("unknown protocol '{}' on type '{}'", d, type_name),
-                    format!("Known protocols: {}", {
-                        let mut sorted = valid; sorted.sort(); sorted.join(", ")
-                    }),
-                    format!("type {}", type_name),
-                ));
-            }
-        }
-    }
-
-    fn register_derive_sigs(&mut self, derives: &[String], type_name: &str) {
-        let type_ty = Ty::Named(type_name.to_string(), vec![]);
-        let value_ty = Ty::Named("Value".to_string(), vec![]);
-        let empty_sb = std::collections::HashMap::new();
-        let empty_pb = std::collections::HashMap::new();
-        for d in derives {
-            match d.as_str() {
-                "Eq" => {
-                    let fn_key = format!("{}.eq", type_name);
-                    if !self.env.functions.contains_key(&fn_key) {
-                        self.env.functions.insert(fn_key, FnSig { params: vec![("a".into(), type_ty.clone()), ("b".into(), type_ty.clone())], ret: Ty::Bool, is_effect: false, generics: vec![], structural_bounds: empty_sb.clone(), protocol_bounds: empty_pb.clone() });
-                    }
-                }
-                "Repr" => {
-                    let fn_key = format!("{}.repr", type_name);
-                    if !self.env.functions.contains_key(&fn_key) {
-                        self.env.functions.insert(fn_key, FnSig { params: vec![("v".into(), type_ty.clone())], ret: Ty::String, is_effect: false, generics: vec![], structural_bounds: empty_sb.clone(), protocol_bounds: empty_pb.clone() });
-                    }
-                }
-                "Codec" => {
-                    let encode_key = format!("{}.encode", type_name);
-                    if !self.env.functions.contains_key(&encode_key) {
-                        self.env.functions.insert(encode_key, FnSig { params: vec![("v".into(), type_ty.clone())], ret: value_ty.clone(), is_effect: false, generics: vec![], structural_bounds: empty_sb.clone(), protocol_bounds: empty_pb.clone() });
-                    }
-                    let decode_key = format!("{}.decode", type_name);
-                    if !self.env.functions.contains_key(&decode_key) {
-                        self.env.functions.insert(decode_key, FnSig { params: vec![("v".into(), value_ty.clone())], ret: Ty::result(type_ty.clone(), Ty::String), is_effect: false, generics: vec![], structural_bounds: empty_sb.clone(), protocol_bounds: empty_pb.clone() });
-                    }
-                }
-                _ => {}
-            }
-        }
-    }
-
-    /// Register a user-defined protocol declaration into env.protocols.
-    fn register_protocol_decl(&mut self, name: &str, generics: &Option<Vec<ast::GenericParam>>, methods: &[ast::ProtocolMethod]) {
-        let gnames: Vec<String> = generics.as_ref()
-            .map(|gs| gs.iter().map(|g| g.name.clone()).collect())
-            .unwrap_or_default();
-
-        // Temporarily register `Self` as a TypeVar so resolve_type_expr handles it
-        self.env.types.insert("Self".to_string(), Ty::TypeVar("Self".to_string()));
-        for gn in &gnames {
-            self.env.types.insert(gn.clone(), Ty::TypeVar(gn.clone()));
-        }
-
-        let method_sigs: Vec<ProtocolMethodSig> = methods.iter().map(|m| {
-            let params: Vec<(String, Ty)> = m.params.iter()
-                .map(|p| (p.name.clone(), self.resolve_type_expr(&p.ty)))
-                .collect();
-            let ret = self.resolve_type_expr(&m.return_type);
-            ProtocolMethodSig {
-                name: m.name.clone(),
-                params,
-                ret,
-                is_effect: m.effect,
-            }
-        }).collect();
-
-        self.env.types.remove("Self");
-        for gn in &gnames {
-            self.env.types.remove(gn);
-        }
-
-        self.env.protocols.insert(name.to_string(), ProtocolDef {
-            name: name.to_string(),
-            generics: gnames,
-            methods: method_sigs,
-        });
-    }
-
-    /// Validate that types declaring `: ProtocolName` have all required convention methods.
-    /// Called after all declarations are registered so all `Type.method` functions are available.
-    fn validate_protocol_impls(&mut self) {
-        // Snapshot type_protocols to avoid borrow conflict
-        let type_protocols: Vec<(String, Vec<String>)> = self.env.type_protocols.iter()
-            .map(|(ty, protos)| (ty.clone(), protos.iter().cloned().collect()))
-            .collect();
-
-        for (type_name, protocol_names) in &type_protocols {
-            for proto_name in protocol_names {
-                let proto_def = match self.env.protocols.get(proto_name) {
-                    Some(p) => p.clone(),
-                    None => continue, // Unknown protocol — already reported by validate_protocols
-                };
-
-                for method_sig in &proto_def.methods {
-                    let fn_key = format!("{}.{}", type_name, method_sig.name);
-                    if !self.env.functions.contains_key(&fn_key) {
-                        // Only emit error for user-defined protocols (not built-in ones
-                        // which may be auto-derived)
-                        let is_builtin = matches!(proto_name.as_str(),
-                            "Eq" | "Repr" | "Ord" | "Hash" | "Codec" | "Encode" | "Decode");
-                        if !is_builtin {
-                            self.emit(err(
-                                format!("type '{}' declares protocol '{}' but missing method '{}'",
-                                    type_name, proto_name, method_sig.name),
-                                format!("Add: fn {}.{}({}) -> {}",
-                                    type_name, method_sig.name,
-                                    method_sig.params.iter()
-                                        .map(|(n, t)| {
-                                            let display_ty = if *t == Ty::TypeVar("Self".to_string()) {
-                                                type_name.clone()
-                                            } else {
-                                                t.display()
-                                            };
-                                            format!("{}: {}", n, display_ty)
-                                        })
-                                        .collect::<Vec<_>>()
-                                        .join(", "),
-                                    {
-                                        let ret = &method_sig.ret;
-                                        if *ret == Ty::TypeVar("Self".to_string()) {
-                                            type_name.clone()
-                                        } else {
-                                            ret.display()
-                                        }
-                                    }),
-                                format!("type {} : {}", type_name, proto_name),
-                            ));
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    fn register_type_decl(&mut self, name: &str, ty: &ast::TypeExpr, deriving: &Option<Vec<String>>,
-                           generics: &Option<Vec<ast::GenericParam>>, prefix: Option<&str>) {
-        if let Some(derives) = deriving {
-            self.validate_protocols(derives, name);
-        }
-        let gnames: Vec<String> = generics.as_ref().map(|gs| gs.iter().map(|g| g.name.clone()).collect()).unwrap_or_default();
-        for gn in &gnames { self.env.types.insert(gn.clone(), Ty::TypeVar(gn.clone())); }
-        let mut resolved = self.resolve_type_expr(ty);
-        for gn in &gnames { self.env.types.remove(gn); }
-        if prefix.is_none() {
-            if let Ty::Variant { name: ref mut vn, ref cases } = resolved {
-                *vn = name.to_string();
-                for case in cases { self.env.constructors.insert(case.name.clone(), (name.to_string(), case.clone())); }
-            }
-        }
-        let key = Self::prefixed_key(prefix, name);
-        self.env.types.insert(key.clone(), resolved.clone());
-        // Also register with unprefixed name so field resolution works
-        // (e.g., g.words where g: KeywordGroup from a module)
-        if prefix.is_some() {
-            self.env.types.insert(name.to_string(), resolved);
-        }
-        if let Some(derives) = deriving {
-            self.register_derive_sigs(derives, name);
-        }
-    }
-
-    fn register_decls(&mut self, decls: &[ast::Decl], prefix: Option<&str>) {
-        for decl in decls {
-            match decl {
-                ast::Decl::Fn { name, params, return_type, effect, r#async, generics, .. } => {
-                    self.register_fn_sig(name, params, return_type, effect, r#async, generics, prefix);
-                }
-                ast::Decl::Type { name, ty, deriving, generics, .. } => {
-                    self.register_type_decl(name, ty, deriving, generics, prefix);
-                    // Track protocol conformances declared via `: ProtocolName`
-                    if let Some(derives) = deriving {
-                        for d in derives {
-                            self.env.type_protocols
-                                .entry(name.clone())
-                                .or_insert_with(std::collections::HashSet::new)
-                                .insert(d.clone());
-                        }
-                    }
-                }
-                ast::Decl::Protocol { name, generics, methods, .. } => {
-                    self.register_protocol_decl(name, generics, methods);
-                }
-                ast::Decl::TopLet { name, ty, value, .. } => {
-                    let rt = ty.as_ref().map(|te| self.resolve_type_expr(te)).unwrap_or_else(|| self.infer_literal_type(value));
-                    let key = Self::prefixed_key(prefix, name);
-                    self.env.top_lets.insert(key, rt);
-                }
-                _ => {}
-            }
-        }
-        // After all decls are registered, validate protocol implementations
-        self.validate_protocol_impls();
     }
 
     // ── Declaration checking ──
